Port cookie conset to BW
