+++ freesound/static/bw-frontend/src/utils/cookies.js
@@ -1,0 +1,25 @@
+export const getCookie = (name) => {
+    let cookieValue = null;
+    if (document.cookie && document.cookie !== '') {
+        const cookies = document.cookie.split(';');
+        for (let i = 0; i < cookies.length; i++) {
+            const cookie = cookies[i].trim();
+            // Does this cookie string begin with the name we want?
+            if (cookie.substring(0, name.length + 1) === (`${name  }=`)) {
+                cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
+                break;
+            }
+        }
+    }
+    return cookieValue;
+}
+
+export const setCookie = (name, value, expiresInDays) => {
+    var expires = "";
+    if (expiresInDays) {
+        var date = new Date();
+        date.setTime(date.getTime() + (expiresInDays*24*60*60*1000));
+        expires = "; expires=" + date.toUTCString();
+    }
+    document.cookie = name + "=" + (value || "")  + expires + "; path=/";
+}
