+++ freesound/static/bw-frontend/src/utils/postRequest.js
@@ -1,20 +1,4 @@
-
-
-const getCookie = (name) => {
-    let cookieValue = null;
-    if (document.cookie && document.cookie !== '') {
-        const cookies = document.cookie.split(';');
-        for (let i = 0; i < cookies.length; i++) {
-            const cookie = cookies[i].trim();
-            // Does this cookie string begin with the name we want?
-            if (cookie.substring(0, name.length + 1) === (`${name  }=`)) {
-                cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
-                break;
-            }
-        }
-    }
-    return cookieValue;
-}
+import { getCookie } from "./cookies";
 
 const makePostRequest = (url, data, successCallback, errorCallback) => {
     const csrftoken = getCookie('csrftoken');
