+++ templates_bw/search/search.html
@@ -285,6 +285,5 @@
         JS bits because we only display the error message{% endcomment %}
         <script src="{% static 'bw-frontend/dist/search.735996ae.js' %}"></script>
         <script src="{% static 'bw-frontend/dist/select.9252c69b.js' %}"></script>
-        <script src="{% static 'bw-frontend/dist/collapsableBlock.10cf5767.js' %}"></script>
     {% endif %}
 {% endblock %}
